requirements.txt
flask
praw
python-dotenv
moviepy
gTTS
openai
groq